+++ Anterminator/AntSim/ControlPlacePheremone.h
@@ -7,7 +7,7 @@ class ControlPlacePheremone
 	float BasePlaceParamater = 1;
 	float ModifiedPlace = 1;
 public:
-	float SpreadSize = 1;
+	int SpreadSize = 1;
 	void Update(World& world, Entity& entity,PheremoneGrid & Pheremone,float s,float v);
 };
